feat: add service & implementation for order
Changes to be committed:
	modified:   app/features/order/Order.scala
	new file:   app/features/order/OrderService.scala
	new file:   app/features/order/OrderServiceImpl.scala
	modified:   app/features/product/ProductServiceImpl.scala

+++ app/features/order/Order.scala
@@ -2,7 +2,7 @@ package features.order
 
 import play.api.libs.json.{Format, Json}
 
-case class Order(id: Int, products: Map[Product, Int])
+case class Order(id: Long, products: Map[Product, Int])
 object Order {
   implicit val format: Format[Order] = Json.format
 }
